requirements.txt
pyfunctional
numpy
pandas
scipy
tqdm
pyinterval